Make ArgGasInfo and ArbAggregator only compile in the test configuration
